+++ bgplot/__init__.py
@@ -1,3 +1,4 @@
 from .features.graphics.graphics import Graphics
-from .features.operations import operations as op
+from .features.operations import operations as ops
 from .core.logger import Logger
+from .features.graphics.core.colors import Colors