+++ bgplot/features/graphics/use_cases/axes_plotting.py
@@ -1,23 +1,36 @@
 from matplotlib.axes import Axes as mplAxes
 
 from ....entities import Axes, Point
+from ..core.colors import Colors
+
+# TMP
+
+
+def _get_color(value: str) -> tuple[float, float, float, float]:
+    stripped_code = value.lstrip('#')
+    rgb_value = tuple(int(stripped_code[i:i+2], 16) for i in (0, 2, 4))
+    r: float = rgb_value[0]/255.0
+    g: float = rgb_value[1]/255.0
+    b: float = rgb_value[2]/255.0
+
+    return (r, g, b)
 
 
 def add_axes(figure: mplAxes, axes: Axes, position: Point, length: float) -> None:
     figure.quiver(position.x, position.y, position.z,
                   axes.x.u, axes.x.v, axes.x.w,
                   length=length,
-                  colors=(1.0, 0.0, 0.0))
+                  colors=_get_color(Colors.red))
 
     figure.quiver(position.x, position.y, position.z,
                   axes.y.u, axes.y.v, axes.y.w,
                   length=length,
-                  colors=(0.0, 1.0, 0.0))
+                  colors=_get_color(Colors.green))
 
     figure.quiver(position.x, position.y, position.z,
                   axes.z.u, axes.z.v, axes.z.w,
                   length=length,
-                  colors=(0.0, 0.0, 1.0))
+                  colors=_get_color(Colors.blue))
 
 
 def add_multiple_axes(figure: mplAxes, axes_list: list[Axes], positions: list[Point], length: float) -> None:
@@ -37,14 +50,14 @@ def add_multiple_axes(figure: mplAxes, axes_list: list[Axes], positions: list[Po
     figure.quiver(x, y, z,
                   u_x, v_x, w_x,
                   length=length,
-                  colors=(1.0, 0.0, 0.0))
+                  colors=_get_color(Colors.red))
 
     figure.quiver(x, y, z,
                   u_y, v_y, w_y,
                   length=length,
-                  colors=(0.0, 1.0, 0.0))
+                  colors=_get_color(Colors.green))
 
     figure.quiver(x, y, z,
                   u_z, v_z, w_z,
                   length=length,
-                  colors=(0.0, 0.0, 1.0))
+                  colors=_get_color(Colors.blue))