+++ bgplot/features/graphics/use_cases/path_plotting.py
@@ -1,0 +1,8 @@
+from collections import deque
+from matplotlib.axes import Axes as mplAxes
+
+from ....entities import Point
+
+
+def add_path(figure: mplAxes, position: Point) -> None:
+    pass