+++ tests/graphics/line_plotting_test.py
@@ -9,13 +9,13 @@ line_2: Line = Line.from_vector_and_point(
     Vector(0.0, 1.0, 1.0), Point(0.0, 0.0, 0.0))
 
 graphics.add_line(line_1)
-graphics.add_line(line_2, line_range=(0.0, 1.0), style='--', color='tab:gray')
+graphics.add_line(line_2, line_range=(0.0, 1.0), style='--', color=bgp.Colors.gray)
 
 graphics.set_title('line')
 graphics.show()
 
 # multiple lines
-graphics.add_lines([line_1, line_2], color='r')
+graphics.add_lines([line_1, line_2], color=bgp.Colors.red)
 
 graphics.set_title('lines')
 graphics.show()